pyproject.toml
[project]
name = "cautious-fiesta"
version = "0.1.0"
description = "A little helper to manage yaml setting files"
authors = [
    {name = "Matteo Bortolotto",email = "yamatteo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dynamic-yaml (>=2.0.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]
license = { text = "MIT" }



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
